pyproject.toml
[tool.poetry]
name = "user-agent-index"
version = "0.1.0"
description = ""
authors = ["Mathieu Lecarme <mathieu@garambrogne.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
tree-sitter = "^0.20.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.292"
pycodestyle = "^2.11.1"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
